Extract drag detection into checkers/drag.py
